Bom atom saves to file
