+++ src/js/molecules/BOM.js
@@ -1,5 +1,6 @@
 import Atom from '../prototypes/atom'
 import GlobalVariables from '../globalvariables'
+import BOMEntry from '../BOM'
 
 
 export default class BillOfMaterials extends Atom{
@@ -12,6 +13,8 @@ export default class BillOfMaterials extends Atom{
         this.name = "Bill Of Materials";
         this.radius = 20;
         
+        this.BOMlist = [];
+        
         this.setValues(values);
     }
     
@@ -20,7 +23,7 @@ export default class BillOfMaterials extends Atom{
         
         var valueList = super.updateSidebar(); //call the super function
         
-        
+        this.createBOM(valueList, this, this.BOMlist);
     }
     
     draw() {
@@ -43,11 +46,53 @@ export default class BillOfMaterials extends Atom{
         return [this.readmeText];
     }
     
+    requestBOM(){
+        //Placeholder
+        return this.BOMlist;
+    }
+    
+    createBOM(list,parent,BOMlist){
+        //aBOMEntry = new bomEntry;
+        
+        
+        list.appendChild(document.createElement('br'));
+        list.appendChild(document.createElement('br'));
+        
+        var div = document.createElement("h3");
+        div.setAttribute("style","text-align:center;");
+        list.appendChild(div);
+        var valueText = document.createTextNode("Bill Of Materials");
+        div.appendChild(valueText);
+        
+        var x = document.createElement("HR");
+        list.appendChild(x);
+        
+        this.BOMlist.forEach(bomItem => {
+            this.createEditableValueListItem(list,bomItem,"BOMitemName", "Item", false)
+            this.createEditableValueListItem(list,bomItem,"numberNeeded", "Number", true)
+            this.createEditableValueListItem(list,bomItem,"costUSD", "Price", true)
+            this.createEditableValueListItem(list,bomItem,"source", "Source", false)
+            var x = document.createElement("HR");
+            list.appendChild(x);
+        });
+        
+        this.createButton(list,parent,"Add BOM Entry",(e) => {
+           this.addBOMEntry();
+        });
+    }
+    
+    addBOMEntry(){
+        console.log("add bom entry ran");
+        this.BOMlist.push(new BOMEntry());
+        
+        this.updateSidebar();
+    }
+    
     serialize(values){
         //Save the readme text to the serial stream
         var valuesObj = super.serialize(values);
         
-        valuesObj.readmeText = this.readmeText;
+        valuesObj.BOMlist = this.BOMlist;
         
         return valuesObj;
